% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_capsule.R
\name{batch_parallel_gsea}
\alias{batch_parallel_gsea}
\title{All-in-One: Batch Parallel GSEA & Report}
\usage{
batch_parallel_gsea(
  gsea_env,
  output_root = "./GSEA_Final_Reports",
  series_name = NULL,
  bidirectional = TRUE,
  workers = 4,
  p_adjust_cutoff = 1,
  top_plots = c(10, 10),
  save_pdf = FALSE,
  dge_list = NULL,
  ...
)
}
\description{
一键全自动：先并行计算，后并行出图，并允许精细控制报告生成参数
}
