% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_capsule.R
\name{batch_plot_gsea}
\alias{batch_plot_gsea}
\title{Batch Parallel GSEA HTML Report Generation}
\usage{
batch_plot_gsea(
  calc_results,
  output_root = "./GSEA_Final_Reports",
  series_name = NULL,
  workers = 4,
  p_adjust_cutoff = 1,
  top_plots = c(20, 20),
  save_pdf = FALSE,
  dge_list = NULL,
  ...
)
}
\description{
接收 batch_calc_gsea 的结果，多核并行生成 HTML 报告和折线图
}
