% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_builder.R
\name{build_custom_msigdb}
\alias{build_custom_msigdb}
\title{Build Custom Mixed MSigDB Dataframe (V10 修复版)}
\usage{
build_custom_msigdb(
  collection_list = list(c("C2", "CP:KEGG_LEGACY"), c("C2", "CP:KEGG_MEDICUS"), c("C5",
    "GO:BP"), c("C5", "GO:CC"), c("C5", "GO:MF")),
  species = "Homo sapiens"
)
}
\description{
Build Custom Mixed MSigDB Dataframe (V10 修复版)
}
