% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/00_class_validations.R

\name{class_validations}

\alias{class_validations}

\title{Core Class Definitions and Validation Functions}

\description{

Define GseaEnv, GseaRes, GseaTask structures and provide internal validation logic.

}

\keyword{internal}

