% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_validations.R
\name{create_gsea_res}
\alias{create_gsea_res}
\title{定义 GseaRes 类}
\usage{
create_gsea_res(
  metadata,
  backend_info,
  contrast_registry,
  de_store,
  expr_bundle,
  geneset_info,
  results
)
}
\description{
GSEA 计算结果胶囊。
}
