% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_validations.R
\name{create_gsea_task}
\alias{create_gsea_task}
\title{Define GseaTask Class}
\usage{
create_gsea_task(gsea_res, meta)
}
\arguments{
\item{gsea_res}{GseaRes object}

\item{meta}{Metadata for the task}
}
\value{
A GseaTask object
}
\description{
Single contrast extraction result object.
}
\examples{
gsea_task <- create_gsea_task(
  gsea_res = NULL,
  meta = data.frame()
)
}
