% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_shiny_app.R
\name{.auto_detect_addition_data}
\alias{.auto_detect_addition_data}
\title{Auto-detect Addition Data}
\usage{
.auto_detect_addition_data()
}
\description{
Search working directory for default addition_data files
}
\keyword{internal}
