% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/01_backends.R

\name{.build_expr_bundle}

\alias{.build_expr_bundle}

\title{Build Expression Data Bundle}

\usage{

.build_expr_bundle(obj, backend)

}

\description{

Uniformly encapsulate expression matrix and metadata

}

\keyword{internal}

