% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/00_class_validations.R

\name{.check_gsea_res}

\alias{.check_gsea_res}

\title{Validate GseaRes Object Integrity}

\usage{

.check_gsea_res(res_obj)

}

\arguments{

\item{res_obj}{GseaRes object to validate}

}

\value{

TRUE if validation passes, otherwise stops with an error

}

\description{

Internal function to ensure GseaRes object contains required components.

}

\keyword{internal}

