% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/utils-accessors.R

\name{.convert_expr_rownames}

\alias{.convert_expr_rownames}

\title{Convert Expression Matrix Row Names to Gene Symbols}

\usage{

.convert_expr_rownames(expr_mat, gene_meta, gsea_res)

}

\arguments{

\item{expr_mat}{An expression matrix.}



\item{gene_meta}{A gene metadata data frame containing both ID and symbol columns.}



\item{gsea_res}{A GseaRes object (used to access backend information).}

}

\value{

The expression matrix with updated row names.

}

\description{

Automatically detect and convert expression matrix row names,

  supporting Ensembl ID to gene symbol conversion.

}

\keyword{internal}

