% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_calculation.R
\name{.enrich_gsea_result}
\alias{.enrich_gsea_result}
\title{GSEA Result Metadata Injection}
\usage{
.enrich_gsea_result(result_df, meta_dict)
}
\arguments{
\item{result_df}{GSEA result dataframe}

\item{meta_dict}{Metadata dictionary}
}
\value{
Enriched dataframe
}
\description{
GSEA Result Metadata Injection
}
\keyword{internal}
