% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/utils-accessors.R

\name{.get_display_symbols}

\alias{.get_display_symbols}

\title{Batch Get Display Gene Symbols}

\usage{

.get_display_symbols(genes, symbol_map)

}

\arguments{

\item{genes}{A character vector of gene identifiers.}



\item{symbol_map}{A case mapping vector generated by \code{.rebuild_symbol_map}.}

}

\value{

A character vector of gene symbols in their original case.

}

\description{

Batch Get Display Gene Symbols

}

\keyword{internal}

