% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/07_shiny_app.R

\name{.validate_addition_data}

\alias{.validate_addition_data}

\title{Validate Addition Data}

\usage{

.validate_addition_data(data)

}

\arguments{

\item{data}{A data.frame to validate}

}

\value{

Validated data.frame with character ID column

}

\description{

Internal function to validate addition_data structure.

  Checks that input is a data.frame with required 'ID' column and

  removes duplicate entries if present.

}

\keyword{internal}

