% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_validations.R
\name{.validate_deseq2_design}
\alias{.validate_deseq2_design}
\title{校验 DESeq2 目标因子}
\usage{
.validate_deseq2_design(dds, target_factor)
}
\arguments{
\item{dds}{DESeqDataSet 对象}

\item{target_factor}{字符串，指定的目标因子}
}
\value{
TRUE 或抛出错误
}
\description{
检查 target_factor 是否存在于 colData 中。
}
\keyword{internal}
