% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/00_class_validations.R

\name{.validate_limma_design}

\alias{.validate_limma_design}

\title{Validate Limma Design Matrix}

\usage{

.validate_limma_design(fit)

}

\arguments{

\item{fit}{MArrayLM object from limma}

}

\value{

TRUE if validation passes, otherwise stops with an error

}

\description{

Enforce no-intercept design (~ 0 + group) to ensure accurate contrast parsing.

}

\keyword{internal}

