% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_gsea_calculation.R
\name{extract_gsea_task_pro}
\alias{extract_gsea_task_pro}
\title{Pro 数据提取器 (文章精细绘图专用)}
\usage{
extract_gsea_task_pro(gsea_capsule, task_name)
}
\arguments{
\item{gsea_capsule}{GseaResPro 计算胶囊}

\item{task_name}{想提取的组别名 (如 "LTA_vs_POSTA")}
}
\value{
DirectionalGSEA 对象。
}
\description{
直接输入对比组名称，提取并包装成原生 DirectionalGSEA 对象，完美兼容底层画图。
}
