% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-codegen.R
\name{generate_summary_code}
\alias{generate_summary_code}
\title{生成 GseaRes 对象摘要代码}
\usage{
generate_summary_code(rds_path = "path/to/your/GSEA_Capsule.rds")
}
\arguments{
\item{rds_path}{字符串，胶囊文件路径。}
}
\value{
字符串，R 代码。
}
\description{
生成用于查看和分析 GseaRes 对象的辅助代码。
}
