% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_shiny_mod_quadrant.R
\name{geom_flat_violin}
\alias{geom_flat_violin}
\title{半小提琴图图层函数}
\usage{
geom_flat_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "dodge",
  trim = TRUE,
  scale = "area",
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
半小提琴图图层函数
}
\keyword{internal}
