% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_shiny_mod_quadrant.R
\docType{data}
\name{GeomFlatViolin}
\alias{GeomFlatViolin}
\title{半小提琴图Geom（为云雨图提供半边小提琴）}
\format{
An object of class \code{GeomFlatViolin} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomFlatViolin
}
\description{
修改自ggplot2 geom_violin源码，仅绘制半边小提琴
}
\keyword{internal}
