% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-accessors.R
\name{get_contrast_registry}
\alias{get_contrast_registry}
\title{Get Contrast Registry}
\usage{
get_contrast_registry(obj)
}
\arguments{
\item{obj}{A GseaEnv or GseaRes object.}
}
\value{
The contrast registry object stored in the input object.
}
\description{
Retrieve the registered contrasts used for differential expression
  analysis, including contrast definitions and metadata.
}
\examples{
\dontrun{
registry <- get_contrast_registry(gsea_res)
names(registry)
}
}
