% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-accessors.R
\name{get_geneset_info}
\alias{get_geneset_info}
\title{Get Gene Set Information}
\usage{
get_geneset_info(obj)
}
\arguments{
\item{obj}{A GseaEnv or GseaRes object.}
}
\value{
An object containing geneset_info with pathway definitions and
  metadata dictionary.
}
\description{
Retrieve gene set collection metadata including pathway definitions
  and annotation dictionary.
}
\examples{
# Load a pre-computed GseaRes shipped with the package
data(precomputed_gseares, package = "GSEAlens")

gsea_res <- precomputed_gseares
gs_info <- get_geneset_info(gsea_res)
names(gs_info)
}
