% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_gsea_environment.R
\name{inspect_gsea_env_pro}
\alias{inspect_gsea_env_pro}
\title{检查计算胶囊 (Pro 引擎)}
\usage{
inspect_gsea_env_pro(env_pro)
}
\arguments{
\item{env_pro}{setup_gsea_env_pro 创建的对象}
}
\description{
检查计算胶囊 (Pro 引擎)
}
