% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_extraction.R
\name{inspect_gsea_res}
\alias{inspect_gsea_res}
\title{View GSEA Result Overview}
\usage{
inspect_gsea_res(gsea_res)
}
\arguments{
\item{gsea_res}{GseaRes object}
}
\value{
Invisibly returns the input object (for chaining).
}
\description{
Print detailed summary of GseaRes object to console.
}
\examples{
\dontrun{
inspect_gsea_res(gsea_res)
}
}
