% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_capsule.R
\name{inspect_gsea_res_pro}
\alias{inspect_gsea_res_pro}
\title{查看 GSEA 计算胶囊结果 (Pro 引擎)}
\usage{
inspect_gsea_res_pro(res_pro)
}
\arguments{
\item{res_pro}{由 batch_calc_gsea_pro() 生成的 DudaliGseaResPro 对象}
}
\description{
格式化打印 DudaliGseaResPro 对象的元数据和各个对比组的富集状态，替代容易导致 RStudio 崩溃的 View()。
}
