% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-i18n.R
\name{list_i18n_keys}
\alias{list_i18n_keys}
\title{获取所有可用翻译键}
\usage{
list_i18n_keys(lang = NULL)
}
\arguments{
\item{lang}{语言代码，默认当前语言。}
}
\value{
字符向量，所有键名（展平格式，使用点号分隔）。
}
\description{
用于开发调试，列出当前语言的所有可用键。
}
