% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_shiny_mod_AI.R
\name{mod_ai_abs_page_server}
\alias{mod_ai_abs_page_server}
\title{AI Interpretation Page Server}
\usage{
mod_ai_abs_page_server(id, gsea_res, data_prep_list, table_controller)
}
\arguments{
\item{id}{Module ID}

\item{gsea_res}{GseaRes object}

\item{data_prep_list}{Data preprocessing list}

\item{table_controller}{Table controller}
}
\description{
AI Interpretation Page Server
}
\keyword{Edge}
\keyword{Leading}
\keyword{internal}
\keyword{判断置信度}
\keyword{基因并格式化为逗号分隔}
\keyword{提取}
