% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_shiny_mod_hubgene.R
\name{mod_hubgene_ui}
\alias{mod_hubgene_ui}
\title{HubGene Network Module UI}
\usage{
mod_hubgene_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Shiny UI tagList
}
\description{
User interface for HubGene Network visualization
}
\keyword{internal}
