% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/12_shiny_mod_multi_plot.R

\name{mod_multi_plot_ui}

\alias{mod_multi_plot_ui}

\title{Combined Pathway Plotting UI}

\usage{

mod_multi_plot_ui(id)

}

\description{

Combined Pathway Plotting UI

}

\keyword{internal}

