% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_shiny_mod_network_edge_detail.R
\name{mod_network_edge_detail_server}
\alias{mod_network_edge_detail_server}
\title{Network Edge Detail Modal Server}
\usage{
mod_network_edge_detail_server(id, parent_session)
}
\arguments{
\item{id}{Module ID}

\item{parent_session}{Parent module session for receiving edge click events}
}
\description{
Network Edge Detail Modal Server
}
\keyword{internal}
