% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/11_shiny_mod_modal.R

\name{mod_pathway_modal_server}

\alias{mod_pathway_modal_server}

\title{Pathway Detail Modal Server}

\usage{

mod_pathway_modal_server(

  id,

  data_prep,

  trigger_event,

  gsea_res,

  pending_genes_reactive,

  update_pending_genes

)

}

\arguments{

\item{id}{Module ID}



\item{data_prep}{Reactive data from data prep module}



\item{trigger_event}{Event that triggers modal opening}



\item{gsea_res}{GseaRes object}



\item{pending_genes_reactive}{Function to get current pending genes}



\item{update_pending_genes}{Function to update pending genes list}

}

\description{

Modal with checkbox selection using standard DT approach.

}

\keyword{internal}

