% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{modDataPrepUI}
\alias{modDataPrepUI}
\title{GSEAlens PRO 模块化重构版 (Shiny Modules架构)}
\usage{
modDataPrepUI(id)
}
\arguments{
\item{id}{模块命名空间ID}
}
\value{
Shiny UI元素列表
}
\description{
修复core_genes作用域、Description换行、基因名大小写匹配、高亮置顶等问题

集成对比组选择、联合绘图控制、表达量类型选择及数据切片功能
}
