% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{modMasterTableServer}
\alias{modMasterTableServer}
\title{主工作台表格Server模块 (修订版)}
\usage{
modMasterTableServer(id, data_prep_module)
}
\arguments{
\item{id}{模块ID}

\item{data_prep_module}{来自模块1的数据流}
}
\value{
响应式向量，包含当前选中的通路ID（用于联合绘图）
}
\description{
渲染交互式DT表格，动态生成HTML按钮，统一命名空间事件监听
}
