% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{modMultiPlotServer}
\alias{modMultiPlotServer}
\title{联合绘图控制台Server模块}
\usage{
modMultiPlotServer(id, data_prep_module, selection_controller)
}
\arguments{
\item{id}{模块ID}

\item{data_prep_module}{来自模块1的数据流（包含plot_subtype和custom_colors）}

\item{selected_ids}{来自模块2的选中通路ID向量（reactive）}
}
\description{
根据模块1的配置和模块2的选中通路执行联合绘图
}
