% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualization.R
\name{plot_directional_gsea}
\alias{plot_directional_gsea}
\title{绘制方向性 GSEA 图}
\usage{
plot_directional_gsea(
  directional_gsea_obj,
  target_pathways,
  main_title = "GSEA Plot",
  subPlot = 3,
  curveCol = NULL,
  add_pval = FALSE,
  show_contrast_in_axis = FALSE,
  ...
)
}
\arguments{
\item{directional_gsea_obj}{封装好的 GseaTask 对象 (由 extract_gsea_task 返回)}

\item{target_pathways}{需要绘制的通路 ID 向量 (支持单个或多个)}

\item{main_title}{主标题名称}

\item{subPlot}{控制 GseaVis 生成的子图数量 (1: 仅富集图; 2: 富集+热图带; 3: 完整带Rank)}

\item{curveCol}{自定义曲线颜色向量}

\item{add_pval}{是否在图上添加统计学标注。默认 FALSE。}

\item{show_contrast_in_axis}{逻辑值。是否在 Rank 图的 x 轴显示对比组信息。
默认为 FALSE，仅在联合画布模式建议设为 TRUE。}

\item{...}{传递给 GseaVis::gseaNb 的额外参数}
}
\value{
ggplot2 复合对象
}
\description{
极客级 GSEA 绘图引擎，完美兼容单通路/多通路合图。
  自动生成优雅图例、拦截并重绘经典红蓝基因分布带。
}
