% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_capsule.R
\name{run_directional_gsea}
\alias{run_directional_gsea}
\title{Run Directional GSEA}
\usage{
run_directional_gsea(
  gsea_env,
  target_id = 1,
  force_left_group = NULL,
  top_n = 40000,
  pvalueCutoff = 1,
  seed = 123
)
}
\description{
Run Directional GSEA
}
