% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_gsea_calculation.R
\name{show_gsea_subsets}
\alias{show_gsea_subsets}
\title{展示 GSEA 胶囊中包含的所有子集 (科学排序)}
\usage{
show_gsea_subsets(meta_dict)
}
\arguments{
\item{meta_dict}{胶囊中的 meta_dict 数据框}
}
\value{
隐式返回排序后的统计数据框
}
\description{
提取并按 H -> C1 -> C2...-> C8 的生物学逻辑严格排序，全量展示子集及其通路数量。
}
