% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-benchmark.R
\name{align_benchmark_data}
\alias{align_benchmark_data}
\title{Align GSEA Performance Data with System Monitoring Data}
\usage{
align_benchmark_data(gsea_res, monitor_csv)
}
\arguments{
\item{gsea_res}{A GseaRes object (must contain metadata$gsea_benchmark)}

\item{monitor_csv}{Path to the CSV file generated by the monitoring script}
}
\value{
A aligned data frame (containing unified time axis fields such as relative_sec)
}
\description{
Aligns timestamps generated by batch_calc_gsea with CSV data from
  an independent monitoring script, producing a time series data frame suitable
  for publication.
}
\examples{
\donttest{
res <- readRDS("GSEA_Capsule_[...].rds")
aligned <- align_benchmark_data(res, "system_monitor.csv")
plot(aligned$relative_sec, aligned$rss_mb, type = "l")
}
}
