% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_capsule.R
\name{batch_calc_gsea}
\alias{batch_calc_gsea}
\title{Batch Parallel GSEA Computation (Only Math)}
\usage{
batch_calc_gsea(
  gsea_env,
  bidirectional = TRUE,
  workers = 25,
  series_name = NULL,
  ...
)
}
\description{
仅进行并行的 GSEA 计算，返回计算结果列表，不出图（速度极快）
}
