% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_capsule.R
\name{batch_parallel_gsea_pro}
\alias{batch_parallel_gsea_pro}
\title{GSEA 全流程一键终极总揽引擎 (All-in-One)}
\usage{
batch_parallel_gsea_pro(
  gsea_env,
  custom_series_name = "Auto_Analysis",
  output_dir = "./GSEA_Output",
  force = FALSE,
  bidirectional = TRUE,
  top_plots = c(15, 15),
  workers = 20,
  dpi = 200,
  minGSSize = 10,
  maxGSSize = 500,
  pvalueCutoff = 1
)
}
\arguments{
\item{gsea_env}{setup_gsea_env_pro 创建的环境胶囊}

\item{custom_series_name}{自定义分析批次名称（如 "2026_03_09_最终汇报"）}

\item{save_rds}{逻辑值，是否自动在根目录留存包含所有信息的数据胶囊}
}
\description{
从环境胶囊一键直达最终的高清 PNG 与自治 HTML 网页，并自动保存复现用的 RDS 文件。
}
