% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/utils-core-genes.R

\name{calculate_overlap_ratio}

\alias{calculate_overlap_ratio}

\title{Calculate ORA Ratio (for DotPlot)}

\usage{

calculate_overlap_ratio(

  pathway_genes,

  de_genes,

  ratio_mode = c("ora", "leading")

)

}

\arguments{

\item{pathway_genes}{Gene set of pathway (defined by TERM2GENE)}



\item{de_genes}{Differentially expressed genes (filtered by p-value threshold)}



\item{ratio_mode}{Character, "ora" (intersection/pathway size) or "leading"

(intersection/DE size)}

}

\value{

Numeric, ratio value

}

\description{

Calculate overlap ratio between pathway genes and differentially

  expressed genes.

}

\examples{

ratio <- calculate_overlap_ratio(c("GAPDH", "TP53", "MYC"), c("TP53", "MYC", "EGFR"))

}

