% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_validations.R
\name{create_gsea_env}
\alias{create_gsea_env}
\title{定义 GseaEnv 类}
\usage{
create_gsea_env(
  backend_info,
  contrast_registry,
  de_store,
  expr_bundle,
  geneset,
  raw_obj
)
}
\description{
标准化的 GSEA 输入环境对象。
}
