% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/00_class_validations.R

\name{create_gsea_res}

\alias{create_gsea_res}

\title{Define GseaRes Class}

\usage{

create_gsea_res(

  metadata,

  backend_info,

  contrast_registry,

  de_store,

  expr_bundle,

  geneset_info,

  results

)

}

\arguments{

\item{metadata}{Metadata information}



\item{backend_info}{Backend information object}



\item{contrast_registry}{Contrast registry}



\item{de_store}{Differential expression store}



\item{expr_bundle}{Expression data bundle}



\item{geneset_info}{Gene set information}



\item{results}{GSEA results}

}

\value{

A GseaRes object

}

\description{

GSEA computation result capsule.

}

\examples{

gsea_res <- create_gsea_res(

  metadata = list(),

  backend_info = list(),

  contrast_registry = data.frame(),

  de_store = data.frame(),

  expr_bundle = list(),

  geneset_info = list(),

  results = list()

)

}

