% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/07_shiny_app.R

\name{.auto_detect_addition_data}

\alias{.auto_detect_addition_data}

\title{Auto-detect Addition Data}

\usage{

.auto_detect_addition_data()

}

\value{

A validated data.frame or NULL if no file found

}

\description{

Search working directory for default addition_data files.

  Checks for 'addition_data_gsealens.rds' or 'addition_data_gsealens.csv'

  in the current working directory.

}

\keyword{internal}

