% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_validations.R
\name{.check_gsea_env}
\alias{.check_gsea_env}
\title{Validate GseaEnv Object Integrity}
\usage{
.check_gsea_env(env_obj)
}
\arguments{
\item{env_obj}{GseaEnv object}
}
\value{
TRUE or stop with error
}
\description{
Internal function to ensure object structure conforms to specification.
}
\keyword{internal}
