% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_validations.R
\name{.check_gsea_env}
\alias{.check_gsea_env}
\title{校验 GseaEnv 对象完整性}
\usage{
.check_gsea_env(env_obj)
}
\arguments{
\item{env_obj}{GseaEnv 对象}
}
\value{
TRUE 或抛出错误
}
\description{
内部函数，确保对象结构符合规范。
}
\keyword{internal}
