% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/01_backends.R

\name{.extract_deseq2_data}

\alias{.extract_deseq2_data}

\title{Extract DESeq2 Data}

\usage{

.extract_deseq2_data(dds, target_factor = NULL)

}

\arguments{

\item{dds}{DESeqDataSet object (must have been processed with DESeq())}



\item{target_factor}{Character. Target factor. If NULL, automatically inferred as the last term in the design formula.}

}

\value{

A list containing contrast_registry, de_store, and expr_bundle

}

\description{

Extract contrast groups and differential analysis results from DESeqDataSet object.

}

\keyword{internal}

