% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/01_backends.R

\name{.extract_limma_data}

\alias{.extract_limma_data}

\title{Extract Limma-Voom Data}

\usage{

.extract_limma_data(fit, expr_data = NULL)

}

\arguments{

\item{fit}{MArrayLM object (must have been processed with eBayes)}



\item{expr_data}{Optional. DGEList or expression matrix.}

}

\value{

A list containing contrast_registry, de_store, and expr_bundle

}

\description{

Extract contrast groups and differential analysis results from MArrayLM object.

}

\keyword{internal}

