% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/utils-accessors.R

\name{.get_display_symbol}

\alias{.get_display_symbol}

\title{Get Display Gene Symbol}

\usage{

.get_display_symbol(gene, symbol_map)

}

\arguments{

\item{gene}{A gene identifier (case-insensitive).}



\item{symbol_map}{A case mapping vector generated by \code{.rebuild_symbol_map}.}

}

\value{

The original case gene symbol, or the input if no mapping exists.

}

\description{

Retrieve the original case gene symbol based on the case mapping.

}

\keyword{internal}

