% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-accessors.R
\name{.get_display_symbols}
\alias{.get_display_symbols}
\title{批量获取显示用基因名}
\usage{
.get_display_symbols(genes, symbol_map)
}
\arguments{
\item{genes}{字符向量}

\item{symbol_map}{大小写映射表}
}
\value{
原始大小写的基因名字符向量
}
\description{
批量获取显示用基因名
}
\keyword{internal}
