% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/vis-gsea-core.R

\name{.gs_info}

\alias{.gs_info}

\title{Internal GSEA Data Extractor}

\usage{

.gs_info(object, geneSetID)

}

\arguments{

\item{object}{A gseaResult object.}



\item{geneSetID}{Character, gene set ID.}

}

\value{

data.frame with runningScore, position, geneList, etc.

}

\description{

Extract running score data from gseaResult object.

  Adapted from GseaVis::gsInfo, using enrichit::gseaScores internally.

}

\keyword{internal}

