% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_shiny_app.R
\name{.launch_gsea_shiny}
\alias{.launch_gsea_shiny}
\title{Launch GSEAlens Shiny App (Internal)}
\usage{
.launch_gsea_shiny(gsea_res, addition_data)
}
\description{
Core Shiny application launcher
}
\keyword{internal}
