% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/07_shiny_app.R

\name{.process_addition_data}

\alias{.process_addition_data}

\title{Process Addition Data}

\usage{

.process_addition_data(addition_data)

}

\arguments{

\item{addition_data}{A data.frame or character vector (file path)}

}

\value{

A validated data.frame with addition data

}

\description{

Internal function to handle various addition_data input types.

  Supports data.frame objects or file paths (.csv, .rds formats).

}

\keyword{internal}

