% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_calculation.R
\name{.process_gsea_chunk}
\alias{.process_gsea_chunk}
\title{Process Single GSEA Chunk}
\usage{
.process_gsea_chunk(
  chunk_task_names,
  task_metadata,
  de_list,
  term2gene,
  meta_dict,
  minGSSize,
  maxGSSize,
  pvalueCutoff
)
}
\arguments{
\item{chunk_task_names}{Character vector}

\item{task_metadata}{Task metadata list}

\item{de_list}{DE data in list form}

\item{term2gene}{Gene set mapping dataframe}

\item{meta_dict}{Metadata dictionary}

\item{minGSSize}{Minimum gene set size}

\item{maxGSSize}{Maximum gene set size}

\item{pvalueCutoff}{P-value threshold}
}
\value{
Result list for current chunk
}
\description{
Process Single GSEA Chunk
}
\keyword{internal}
