% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-accessors.R
\name{.process_sample_meta}
\alias{.process_sample_meta}
\title{Internal Sample Metadata Processing}
\usage{
.process_sample_meta(sample_meta, expr_bundle)
}
\description{
Unified sample metadata processing: fixes row names, unifies group column names.
}
\keyword{internal}
