% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-accessors.R
\name{.process_sample_meta}
\alias{.process_sample_meta}
\title{Internal Sample Metadata Processing}
\usage{
.process_sample_meta(sample_meta, expr_bundle)
}
\description{
Internal Sample Metadata Processing
}
\keyword{internal}
