% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-accessors.R
\name{.rebuild_symbol_map}
\alias{.rebuild_symbol_map}
\title{按需重建大小写映射表}
\usage{
.rebuild_symbol_map(gsea_res, contrast_id)
}
\arguments{
\item{gsea_res}{GseaRes 对象}

\item{contrast_id}{对比组 ID}
}
\value{
命名向量，names = 大写基因名，values = 原始大小写
}
\description{
从 de_store 中重建 gene_symbol 大小写映射，用于显示层
}
\keyword{internal}
