% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/01_backends.R

\name{.standardize_de_columns}

\alias{.standardize_de_columns}

\title{Standardize Differential Expression Table Column Names}

\usage{

.standardize_de_columns(df, backend)

}

\description{

Unify column names from different backends to: gene_symbol, logFC, stat, pvalue, padj

}

\keyword{internal}

