% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_shiny_app.R
\name{.validate_addition_data}
\alias{.validate_addition_data}
\title{Validate Addition Data}
\usage{
.validate_addition_data(data)
}
\description{
Internal function to validate addition_data structure
}
\keyword{internal}
