% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/00_class_validations.R

\name{.validate_deseq2_design}

\alias{.validate_deseq2_design}

\title{Validate DESeq2 Target Factor}

\usage{

.validate_deseq2_design(dds, target_factor)

}

\arguments{

\item{dds}{DESeqDataSet object}



\item{target_factor}{String, the target factor to validate}

}

\value{

TRUE if validation passes, otherwise stops with an error

}

\description{

Check if target_factor exists in colData and is properly formatted.

}

\keyword{internal}

