% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/05_extraction.R

\name{extract_gsea_task}

\alias{extract_gsea_task}

\title{Extract GSEA Results for Specific Contrast}

\usage{

extract_gsea_task(gsea_res, contrast_id, target_collection = "ALL")

}

\arguments{

\item{gsea_res}{GseaRes object (generated by batch_calc_gsea)}



\item{contrast_id}{String. Contrast group ID (e.g., "A_vs_B").}



\item{target_collection}{String vector. Gene set collection to extract

(e.g., "H", "C2:CP:KEGG_LEGACY"). Default "ALL" means extract all.}

}

\value{

GseaTask object.

}

\description{

Extract results for a specified contrast from a GseaRes capsule.

  Supports subsetting by gene set collection and automatically recalculates FDR.

}

\examples{

# Load a pre-computed GseaRes shipped with the package

gsea_res <- readRDS(system.file(

  "extdata", "precomputed_gseares.rds", package = "GSEAlens"

))

task <- extract_gsea_task(gsea_res, contrast_id = "untrt_vs_trt")

class(task)

}

