% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/06_visualization.R

\name{generate_gsea_html_report}

\alias{generate_gsea_html_report}

\title{Generate GSEA HTML Report}

\usage{

generate_gsea_html_report(

  res_obj,

  output_base_dir = NULL,

  p_adjust_cutoff = 1,

  top_plots = c(15, 15),

  dpi = 200

)

}

\arguments{

\item{res_obj}{GseaTask object}



\item{output_base_dir}{Output folder path. If left empty (NULL), it will automatically follow the original project location of the calculation capsule.}



\item{p_adjust_cutoff}{FDR filtering threshold, default 1}



\item{top_plots}{Number of pathways to plot detailed sub-pages, format c(positive_count, negative_count)}



\item{dpi}{Resolution for GSEA enrichment plots, default 200}

}

\value{

Path to the generated HTML report (invisible)

}

\description{

Automatically avoids column name contamination, intelligently extracts expression matrices from objects,

and outputs comfortable web reports with dual p-value display and heatmaps.

Uses the ComplexHeatmap engine to draw high-quality heatmaps, perfectly replicating pheatmap aesthetics.

}

\examples{

\dontrun{

generate_gsea_html_report(res_obj, output_base_dir = tempdir())

}

}

