% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_code_generator.R
\name{generate_hubgene_code}
\alias{generate_hubgene_code}
\title{Generate Publication HubGene Network Code}
\usage{
generate_hubgene_code(
  pathway_nodes,
  gene_nodes,
  edge_df,
  pw_size_mode = "setsize",
  seed = 42L,
  base_size = 12
)
}
\arguments{
\item{pathway_nodes}{data.frame: id, NES, FDR, setSize.}

\item{gene_nodes}{data.frame: id, stat, degree (optional column le_count).}

\item{edge_df}{data.frame: from, to.}

\item{pw_size_mode}{Character. One of "fixed", "fdr", "setsize".}

\item{seed}{Integer.}

\item{base_size}{Numeric.}
}
\value{
Character scalar of executable R code.
}
\description{
Return an executable R script reproducing the HubGene module's
  visNetwork view as a static ggplot2 figure. Node size is controlled by
  the same three-mode logic (fixed / fdr / setsize) as the interactive view.
}

\examples{

# Construct minimal pathway/gene/edge data.frames inline.

mini_pw <- data.frame(

  id = c("PATHWAY_A"), NES = c(1.5), FDR = c(0.01), setSize = c(50L)

)

mini_gn <- data.frame(

  id = c("GENE_X", "GENE_Y"), stat = c(2.5, -1.8), degree = c(3L, 5L)

)

mini_ed <- data.frame(

  from = c("PATHWAY_A", "PATHWAY_A"),

  to   = c("GENE_X", "GENE_Y")

)

code <- generate_hubgene_code(mini_pw, mini_gn, mini_ed)

cat(code)

}
