% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/utils_hubgene.R

\name{generate_hubgene_hover_text}

\alias{generate_hubgene_hover_text}

\title{Generate HubGene Hover Text}

\usage{

generate_hubgene_hover_text(

  node,

  node_type,

  left_group = "A",

  right_group = "B"

)

}

\arguments{

\item{node}{Node data row (as list with named elements)}



\item{node_type}{Node type: "pathway" or "gene"}



\item{left_group}{Left group name}



\item{right_group}{Right group name}

}

\value{

Character string with HTML content for hover tooltip

}

\description{

Generate HTML text for plotly hover tooltips.

}

\examples{

# Example with a pathway node

pathway_node <- list(

  id = "HALLMARK_INFLAMMATORY_RESPONSE",

  NES = 2.3, FDR = 0.001, pvalue = 0.0005,

  n_core = 15, n_total = 50

)

hover_pathway <- generate_hubgene_hover_text(pathway_node, "pathway")

cat(hover_pathway)



# Example with a gene node

gene_node <- list(id = "GENE_A", log2FC = 1.8)

hover_gene <- generate_hubgene_hover_text(gene_node, "gene")

cat(hover_gene)

}

