% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/15_code_generator.R

\name{generate_joint_canvas_code}

\alias{generate_joint_canvas_code}

\title{Generate Joint Canvas Code}

\usage{

generate_joint_canvas_code(

  GSEAlens_res,

  contrast_ids,

  target_pathways,

  ncol = 2

)

}

\arguments{

\item{GSEAlens_res}{GseaRes object}



\item{contrast_ids}{Character vector. Multiple contrast IDs}



\item{target_pathways}{Character vector. Pathway IDs}



\item{ncol}{Integer. Number of columns}

}

\value{

A character string containing executable R code

}

\description{

Generate Joint Canvas Code

}

\examples{

\dontrun{

code <- generate_joint_canvas_code(GSEAlens_res = gsea_res,

                                   contrast_ids = c("A_vs_B"),

                                   target_pathways = c("HALLMARK_INFLAMMATORY_RESPONSE"))

cat(code)

}

}

