% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_code_generator.R
\name{generate_joint_canvas_code}
\alias{generate_joint_canvas_code}
\title{Generate Joint Canvas Code}
\usage{
generate_joint_canvas_code(
  GSEAlens_res,
  contrast_ids,
  target_pathways,
  ncol = 2
)
}
\arguments{
\item{GSEAlens_res}{GseaRes object}

\item{contrast_ids}{Character vector. Multiple contrast IDs}

\item{target_pathways}{Character vector. Pathway IDs to plot}

\item{ncol}{Integer. Number of columns for layout}
}
\value{
A character string containing executable R code
}
\description{
Generate code for the Joint GSEA Canvas (Tab 4)
}
