% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_code_generator.R
\name{generate_pathway_plot_code}
\alias{generate_pathway_plot_code}
\title{Generate Pathway Plotting Code}
\usage{
generate_pathway_plot_code(
  GSEAlens_res,
  contrast_id,
  target_pathways,
  user_genes = character(0),
  expr_type = "logcpm"
)
}
\arguments{
\item{GSEAlens_res}{GseaRes object}

\item{contrast_id}{Character. Contrast ID for the analysis}

\item{target_pathways}{Character vector. Pathway IDs to plot}

\item{user_genes}{Character vector. Genes of interest to highlight}

\item{expr_type}{Character. Expression value type}
}
\value{
A character string containing executable R code
}
\description{
Generate ready-to-run R code for GSEA pathway visualization.
}
