% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-core-genes.R
\name{get_core_genes_for_pathway}
\alias{get_core_genes_for_pathway}
\title{Extract Pathway Core Genes (Leading Edge)}
\usage{
get_core_genes_for_pathway(gsea_res_obj, pathway_id)
}
\arguments{
\item{gsea_res_obj}{GseaRes object or single GSEA result (list(status=, data=, ...))}

\item{pathway_id}{Character, pathway ID}
}
\value{
Character vector, core genes (uppercase normalized)
}
\description{
Parse the core_enrichment field from GSEA results to extract gene list.
  This is the fundamental data interface for all visualization modules in 03/04.
}
\examples{
\dontrun{
core_genes <- get_core_genes_for_pathway(gsea_res, "HALLMARK_OXIDATIVE_PHOSPHORYLATION")
}
}
