% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/utils-core-genes.R

\name{get_core_genes_list}

\alias{get_core_genes_list}

\title{Batch Extract Core Genes for Multiple Pathways}

\usage{

get_core_genes_list(gsea_task_obj, pathway_ids)

}

\arguments{

\item{gsea_task_obj}{GseaTask object (single contrast)}



\item{pathway_ids}{Character vector, pathway ID list}

}

\value{

Named list, names = pathway_id, values = core_genes vector

}

\description{

Batch data preparation for 03 module Network/UpSet/Chord visualizations.

}

\examples{

# Load a pre-computed GseaRes and extract a task

gsea_res <- readRDS(system.file(

  "extdata", "precomputed_gseares.rds", package = "GSEAlens"

))

task <- extract_gsea_task(gsea_res, contrast_id = "untrt_vs_trt")

res_df <- as.data.frame(task$gsea_res)

pw_ids <- head(res_df$ID, 3)

genes_list <- get_core_genes_list(task, pw_ids)

length(genes_list)

}

