% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-core-genes.R
\name{get_core_genes_list}
\alias{get_core_genes_list}
\title{批量提取多个通路的Core Genes}
\usage{
get_core_genes_list(gsea_task_obj, pathway_ids)
}
\arguments{
\item{gsea_task_obj}{GseaTask对象（单个对比组）}

\item{pathway_ids}{字符向量，通路ID列表}
}
\value{
命名列表，names=pathway_id, values=core_genes向量
}
\description{
为03模块的Network/UpSet/Chord提供批量数据准备
}
