% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-accessors.R
\name{get_de_table}
\alias{get_de_table}
\title{获取差异分析表}
\usage{
get_de_table(obj, contrast_id)
}
\arguments{
\item{obj}{GseaEnv 或 GseaRes 对象}

\item{contrast_id}{字符串。对比组 ID (如 "A_vs_B")。}
}
\value{
data.frame (包含 gene_symbol, logFC, stat, pvalue, padj)
}
\description{
从 de_store 中提取指定对比组的结果。
}
