% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-accessors.R
\name{get_de_table}
\alias{get_de_table}
\title{Get Differential Expression Table}
\usage{
get_de_table(obj, contrast_id)
}
\arguments{
\item{obj}{GseaEnv or GseaRes object}

\item{contrast_id}{Character. Contrast ID (e.g., "A_vs_B").}
}
\value{
data.frame (containing gene_symbol, logFC, stat, pvalue, padj)
}
\description{
Extract results for a specific contrast from de_store.
}
