% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-accessors.R
\name{get_expr_matrix}
\alias{get_expr_matrix}
\title{Get Expression Matrix}
\usage{
get_expr_matrix(obj, type = "default", ...)
}
\arguments{
\item{obj}{GseaEnv or GseaRes object}

\item{type}{Character. Options: "default" (default display), "raw", "cpm", "logcpm", "vst", "fpkm", "tpm".}

\item{...}{Additional arguments}
}
\value{
Expression matrix (genes x samples)
}
\description{
Extract expression matrix from GseaEnv or GseaRes, supporting multiple normalization methods.
}
