% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/utils_hubgene.R

\name{get_hubgene_legend}

\alias{get_hubgene_legend}

\title{Get Color Legend Configuration}

\usage{

get_hubgene_legend(left_group = "A", right_group = "B")

}

\arguments{

\item{left_group}{Left group name (e.g., treatment or case)}



\item{right_group}{Right group name (e.g., control or reference)}

}

\value{

A named list containing:

  \item{pathway_up}{Color for up-regulated pathways}

  \item{pathway_down}{Color for down-regulated pathways}

  \item{gene_up}{Color for up-regulated genes}

  \item{gene_down}{Color for down-regulated genes}

  \item{labels}{Named list of legend labels}

}

\description{

Generate color legend data for HubGene network plot.

}

\examples{

legend <- get_hubgene_legend("Control", "Treatment")

}

