% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/utils-core-genes.R

\name{get_term_genes}

\alias{get_term_genes}

\title{Get Full Pathway Gene Set (TERM2GENE)}

\usage{

get_term_genes(gsea_res, pathway_id)

}

\arguments{

\item{gsea_res}{GseaRes object}



\item{pathway_id}{Pathway ID}

}

\value{

Character vector

}

\description{

Extract complete gene list for a pathway from geneset_info.

}

\examples{

# Load a pre-computed GseaRes shipped with the package

gsea_res <- readRDS(system.file(

  "extdata", "precomputed_gseares.rds", package = "GSEAlens"

))

pw_id <- gsea_res$geneset_info$term2gene$gs_name[1]

genes <- get_term_genes(gsea_res, pw_id)

head(genes)

}

