% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/02_setup.R

\name{inspect_gsea_env}

\alias{inspect_gsea_env}

\title{View GSEA Environment Object Overview}

\usage{

inspect_gsea_env(env_obj)

}

\arguments{

\item{env_obj}{GseaEnv object}

}

\value{

Invisibly returns the input object (for chaining).

}

\description{

Print detailed information about the GseaEnv object to the console, including backend type,

contrast list, and gene set status.

}

\examples{

# Build a GseaEnv from pre-computed inputs shipped with the package

precomp <- readRDS(system.file(

  "extdata", "preprocessed_limma.rds", package = "GSEAlens"

))

pathways <- readRDS(system.file(

  "extdata", "gsea_pathwaysets_toy.rds", package = "GSEAlens"

))

env <- setup_gsea_env(

  fit       = precomp$fit,

  pathway_obj = pathways,

  expr_data = precomp$gsea_limma_voom_data

)

inspect_gsea_env(env)

}

