% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_builder.R
\name{inspect_gsea_env}
\alias{inspect_gsea_env}
\title{检查并打印 GSEA 环境状态}
\usage{
inspect_gsea_env(gsea_env)
}
\arguments{
\item{gsea_env}{由 setup_gsea_env() 创建的对象}
}
\description{
提供一个精美格式化的控制台面板，展示胶囊内部状态。
}
