% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_capsule.R
\name{inspect_gsea_env}
\alias{inspect_gsea_env}
\title{Inspect GSEA Environment}
\usage{
inspect_gsea_env(gsea_env)
}
\arguments{
\item{gsea_env}{A GseaEnv object created by setup_gsea_env()}
}
\description{
Provides a beautifully formatted summary of the GseaEnv object, guiding the user on how to proceed to the next step.
}
