% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_setup.R
\name{inspect_gsea_env}
\alias{inspect_gsea_env}
\title{View GSEA Environment Object Overview}
\usage{
inspect_gsea_env(env_obj)
}
\arguments{
\item{env_obj}{GseaEnv object}
}
\description{
Print detailed information about the GseaEnv object to the console, including backend type,
contrast list, and gene set status.
}
\examples{
if(interactive()){
# Placeholder for function example
}
}
