% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_capsule.R
\name{inspect_gsea_env_pro}
\alias{inspect_gsea_env_pro}
\title{检查计算胶囊 (Pro 引擎)}
\usage{
inspect_gsea_env_pro(env_pro)
}
\arguments{
\item{env_pro}{setup_gsea_env_pro 创建的对象}
}
\description{
极其美观地打印当前工作流所装载的所有元数据和对比任务。
}
