% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/05_extraction.R

\name{inspect_gsea_res}

\alias{inspect_gsea_res}

\title{View GSEA Result Overview}

\usage{

inspect_gsea_res(gsea_res)

}

\arguments{

\item{gsea_res}{GseaRes object}

}

\value{

Invisibly returns the input object (for chaining).

}

\description{

Print detailed summary of GseaRes object to console.

}

\examples{

# Load a pre-computed GseaRes shipped with the package

gsea_res <- readRDS(system.file(

  "extdata", "precomputed_gseares.rds", package = "GSEAlens"

))

inspect_gsea_res(gsea_res)

}

