% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_gsea_calculation.R
\name{inspect_gsea_res_pro}
\alias{inspect_gsea_res_pro}
\title{查看 GSEA 计算胶囊结果}
\usage{
inspect_gsea_res_pro(res_pro)
}
\arguments{
\item{res_pro}{batch_calc_gsea_pro() 生成的对象}
}
\description{
格式化打印 GseaResPro 对象的元数据和各个对比组的富集状态，避免直接 View 导致卡顿。
}
