% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{launch_gsea_app}
\alias{launch_gsea_app}
\title{启动 GSEAlens PRO 2.0 终极全息探索空间}
\usage{
launch_gsea_app(res_capsule)
}
\arguments{
\item{res_capsule}{您环境中的 GseaResPro 对象}
}
\description{
彻底终结 \link{object Object} 报错、修复 ALL 选项、加入 Leading Edge 追踪及宽屏 Dashboard。
}
