% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{launch_gsea_app}
\alias{launch_gsea_app}
\title{启动 GSEAlens PRO 2.0 终极全息探索空间}
\usage{
launch_gsea_app(res_capsule)
}
\arguments{
\item{res_capsule}{环境中的 GseaResPro 对象}
}
\description{
强制洗血换骨，动态切片并合并子集合，重算 FDR。无缝四重联动，散点图全局组别展示。
}
