% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/17_shiny_mod_AI.R

\name{mod_ai_abs_page_ui}

\alias{mod_ai_abs_page_ui}

\title{AI Interpretation Page UI}

\usage{

mod_ai_abs_page_ui(id)

}

\arguments{

\item{id}{Module ID}

}

\description{

AI Interpretation Page UI

}

\keyword{internal}

