% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_shiny_mod_data_prep.R
\name{mod_data_prep_server}
\alias{mod_data_prep_server}
\title{Data Preprocessing Module Server (Joint Canvas Parameter Passing Version)}
\usage{
mod_data_prep_server(id, gsea_res)
}
\description{
Data Preprocessing Module Server (Joint Canvas Parameter Passing Version)
}
\keyword{internal}
