% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_shiny_mod_hubgene_vis.R
\name{mod_hubgene_vis_ui}
\alias{mod_hubgene_vis_ui}
\title{HubGene Network (visNetwork) Module UI}
\usage{
mod_hubgene_vis_ui(id)
}
\description{
HubGene Network (visNetwork) Module UI
}
\keyword{internal}
