% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/16_shiny_mod_hubgene_vis.R

\name{mod_hubgene_vis_ui}

\alias{mod_hubgene_vis_ui}

\title{HubGene Network Module UI}

\usage{

mod_hubgene_vis_ui(id)

}

\arguments{

\item{id}{Character string used to namespace the module.}

}

\value{

A Shiny tagList containing the module's UI elements.

}

\description{

Shiny module UI for visualizing HubGene networks using visNetwork.

Provides interactive controls for pathway selection modes, physics

simulation parameters, node sizing, and network statistics display.

}

\keyword{internal}

