% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_shiny_mod_joint_canvas.R
\name{mod_joint_canvas_server}
\alias{mod_joint_canvas_server}
\title{联合GSEA填充画布模块 Server（侧边栏控制版）}
\usage{
mod_joint_canvas_server(id, gsea_res, data_prep_list, table_result)
}
\description{
1. 接收侧边栏控制参数（排列模式）
  2. 无最大行数限制
  3. 美学与multi_plot完全一致
}
\keyword{internal}
