% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_shiny_mod_joint_canvas.R
\name{mod_joint_canvas_server}
\alias{mod_joint_canvas_server}
\title{Joint GSEA Fill Canvas Module Server (Sidebar Control Version)}
\usage{
mod_joint_canvas_server(id, gsea_res, data_prep_list, table_result)
}
\description{
1. Receive sidebar control parameters (arrangement mode)
  2. No maximum row limit
  3. Aesthetic consistency with multi_plot
}
\keyword{internal}
