% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_shiny_mod_table.R
\name{mod_master_table_server}
\alias{mod_master_table_server}
\title{Master Workspace Table Server}
\usage{
mod_master_table_server(id, data_prep)
}
\arguments{
\item{id}{Module ID}

\item{data_prep}{Reactive data from the data preprocessing module}
}
\value{
List containing selected_pathways and show_modal
}
\description{
Provides data display, CSV annotation merging, and column display control with decoupled checkbox state.
}
\keyword{internal}
