% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_shiny_mod_table.R
\name{mod_master_table_server}
\alias{mod_master_table_server}
\title{主工作台表格 Server}
\usage{
mod_master_table_server(id, data_prep)
}
\arguments{
\item{id}{模块 ID}

\item{data_prep}{来自数据预处理模块的响应式数据}
}
\value{
列表，包含 selected_pathways (reactive) 和 show_modal (reactive)
}
\description{
主工作台表格 Server
}
\keyword{internal}
