% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/09_shiny_mod_table.R

\name{mod_master_table_server}

\alias{mod_master_table_server}

\title{Master Workspace Table Server}

\usage{

mod_master_table_server(id, data_prep, addition_data = NULL)

}

\arguments{

\item{id}{Module ID}



\item{data_prep}{Reactive data from the data preprocessing module}



\item{addition_data}{Optional. A data frame with pathway annotations to merge.}

}

\value{

List containing selected_pathways, show_modal, and selection control functions

}

\description{

Provides data display, column display control with decoupled checkbox state.

  Supports merging addition_data columns into the main table.

  Includes external selection update capabilities for bidirectional sync.

}

\keyword{internal}

