% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/09_shiny_mod_table.R

\name{mod_master_table_ui}

\alias{mod_master_table_ui}

\title{Master Workspace Table UI}

\usage{

mod_master_table_ui(id)

}

\arguments{

\item{id}{Module ID}

}

\value{

Shiny UI tagList

}

\description{

Interactive data table module for displaying GSEA pathway results

  with checkbox selection and modal integration.

}

\keyword{internal}

