% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_shiny_mod_table.R
\name{mod_master_table_ui}
\alias{mod_master_table_ui}
\title{Master Workspace Table UI (Phase 11 Complete + Checkbox Decoupling)}
\usage{
mod_master_table_ui(id)
}
\description{
Master Workspace Table UI (Phase 11 Complete + Checkbox Decoupling)
}
\keyword{internal}
