% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_shiny_mod_multi_plot.R
\name{mod_multi_plot_server}
\alias{mod_multi_plot_server}
\title{Combined Pathway Plotting Server}
\usage{
mod_multi_plot_server(id, data_prep, table_controller, gsea_res = NULL)
}
\description{
Combined Pathway Plotting Server
}
\keyword{internal}
