% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_shiny_mod_modal.R
\name{mod_pathway_modal_server}
\alias{mod_pathway_modal_server}
\title{详情弹窗 Server}
\usage{
mod_pathway_modal_server(id, data_prep, trigger_event, gsea_res)
}
\arguments{
\item{id}{模块 ID}

\item{data_prep}{数据流}

\item{trigger_event}{来自表格的点击事件 (reactive)}

\item{gsea_res}{GseaRes 对象 (用于获取表达矩阵)}
}
\description{
ComplexHeatmap重构：Leading Edge间隙、动态排序、无高度限制、基因名颜色区分
}
\keyword{internal}
