% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_shiny_mod_modal.R
\name{mod_pathway_modal_server}
\alias{mod_pathway_modal_server}
\title{Pathway Detail Modal Server - Fixed CPM Logic}
\usage{
mod_pathway_modal_server(id, data_prep, trigger_event, gsea_res)
}
\description{
Forces raw CPM usage regardless of expression_type selection,
  detects and reverses log2 transformation if necessary, displays pathway name in title
}
\keyword{internal}
