% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_shiny_mod_pathway_relation.R
\name{mod_pathway_relation_server}
\alias{mod_pathway_relation_server}
\title{Pathway Relationship Exploration Module Server}
\usage{
mod_pathway_relation_server(id, data_prep_list, gsea_res, table_result = NULL)
}
\description{
Dual-mode pathway network visualization with defensive rendering
}
\keyword{internal}
