% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_shiny_mod_pathway_relation.R
\name{mod_pathway_relation_ui}
\alias{mod_pathway_relation_ui}
\title{通路关系探索模块 UI}
\usage{
mod_pathway_relation_ui(id)
}
\description{
包含DotPlot、Chord、Network、UpSet四个子Tab
}
\keyword{internal}
