% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/13_shiny_mod_pathway_relation.R

\name{mod_pathway_relation_ui}

\alias{mod_pathway_relation_ui}

\title{Pathway Relationship Exploration Module UI}

\usage{

mod_pathway_relation_ui(id)

}

\description{

Network visualization module with dual-mode pathway selection

}

\keyword{internal}

