% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_shiny_mod_quadrant.R
\name{mod_quadrant_server}
\alias{mod_quadrant_server}
\title{Quadrant Linkage Server (Multi-select + Sort Fix)}
\usage{
mod_quadrant_server(id, data_prep_list, gsea_res)
}
\description{
Supports continuous clicking to mark multiple pathways, Boxplot sorting forced lock
}
\keyword{internal}
