% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_shiny_mod_quadrant.R
\name{mod_quadrant_server}
\alias{mod_quadrant_server}
\title{Quadrant Linkage Server}
\usage{
mod_quadrant_server(id, data_prep_list, gsea_res)
}
\arguments{
\item{id}{Module ID}

\item{data_prep_list}{List from data prep module}

\item{gsea_res}{GseaRes object}
}
\description{
Server logic with unified gene pool (applied_genes)
}
\keyword{internal}
