% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_capsule.R
\name{plot_directional_gsea}
\alias{plot_directional_gsea}
\title{Plot Directional GSEA (Bulletproof & Restores Classic Red/Blue Gradient)}
\usage{
plot_directional_gsea(
  directional_gsea_obj,
  target_pathways,
  curveCol = NULL,
  main_title = "GSEA Plot",
  subPlot = 3,
  ...
)
}
\arguments{
\item{directional_gsea_obj}{封装好的 DirectionalGSEA 对象 (需包含 gsea_res 和 meta)}

\item{target_pathways}{需要绘制的通路 ID 向量 (如 c("HALLMARK_...", ...))}

\item{curveCol}{自定义曲线颜色向量 (可选)}

\item{main_title}{主标题名称}

\item{subPlot}{控制 GseaVis::gseaNb 生成的子图数量。
1: 仅GSEA富集图; 2: GSEA富集图 + 基因表达热图; 3: GSEA富集图 + 热图 + Rank List。}

\item{...}{传递给 GseaVis::gseaNb 的额外参数。}
}
\value{
一个 ggplot2 对象，表示 GSEA 结果图。
}
\description{
稳健的 GSEA 绘图函数，完美处理多通路合图、自动生成优雅的图例名称、
并自带经典红蓝底部分布图。此版本通过 \code{subPlot} 参数控制是否显示 Rank List。
}
