% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/utils-benchmark.R

\name{plot_gsea_memory}

\alias{plot_gsea_memory}

\title{Plot GSEA Benchmark Memory Curve}

\usage{

plot_gsea_memory(aligned_data, highlight_phases = TRUE)

}

\arguments{

\item{aligned_data}{Aligned data frame returned by align_benchmark_data}



\item{highlight_phases}{Whether to color by phase, default TRUE}

}

\value{

A ggplot object

}

\description{

Generates a standard memory usage curve plot for publication.

}

\examples{

\dontrun{

data <- align_benchmark_data(gsea_res, "system_monitor.csv")

plot_gsea_memory(data)

plot_gsea_memory(data, highlight_phases = FALSE)

}

}

