% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-benchmark.R
\name{plot_gsea_memory}
\alias{plot_gsea_memory}
\title{绘制GSEA基准测试内存曲线}
\usage{
plot_gsea_memory(aligned_data, highlight_phases = TRUE)
}
\arguments{
\item{aligned_data}{align_benchmark_data返回的对齐数据框}

\item{highlight_phases}{是否按阶段着色，默认TRUE}
}
\value{
ggplot对象
}
\description{
为发表文章生成标准内存使用曲线图
}
