% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-benchmark.R
\name{print.GseaBenchmarkAligned}
\alias{print.GseaBenchmarkAligned}
\title{Print GSEA Benchmark Alignment Results}
\usage{
\method{print}{GseaBenchmarkAligned}(x, ...)
}
\description{
Print GSEA Benchmark Alignment Results
}
