% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_capsule.R
\name{setup_gsea_env}
\alias{setup_gsea_env}
\title{Setup GSEA Environment (Polymorphic Engine)}
\usage{
setup_gsea_env(
  fit,
  custom_df = NULL,
  gmt_file = NULL,
  custom_set_name = "Default_Set",
  category = "H",
  subcategory = NULL,
  species = "Homo sapiens"
)
}
\description{
Setup GSEA Environment (Polymorphic Engine)
}
