% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/utils-accessors.R

\name{utils-accessors}

\alias{utils-accessors}

\title{Data Accessor Utilities}

\description{

Provides unified data extraction interface, isolating underlying

  object structure differences. Fixes DESeq2 sample name matching issues.

}

\keyword{internal}

